TMS-89, 118 JUnit for ManagerView and fixed the issues in the backend code of Manager Edit Member and added the Assignmenet.
